trs::lexicon: revert to the inconsistent typing state
We have at least three options for how we treat typed rules:
1. switch over to globally valid rules only (i.e. `v0_ v1_ = v1_` is invalid).
2. switch over to rewriting guarded by type compatibility (i.e. `v0_ v1_ = v1_`
is valid but only applies when the types of the term and rule are compatible).
3. adopt a slightly more inconsistent state in which we allow rules which are
only locally valid but apply them globally, trusting that bad rules will be
weeded out via the likelihood (i.e. `v0_ v1_ = v1_` is valid and always applies,
even if it leaves the term in an inconsistent state).

This commit reverts from 1 to 3, as that's worked well so far. We'll target 1
or 2 in the future.